Lots of app structure and testing improvements: - Finished initial GET /notifications implementation, plus tests - Added scripts for loading sample data - Added fishery module for test factories - Module security updates - README updates

+++ api-v2/test/test-helpers.ts
@@ -1,7 +1,9 @@
 import { Model } from 'mongoose';
 
-export async function truncateAllCollections(anyModel: Model<any>) {
+export const truncateAllCollections = async (anyModel: Model<any>) => {
   for (const collection of Object.values(anyModel.db.collections)) {
-    await collection.deleteMany({});
+    if (collection) {
+      await collection.deleteMany({});
+    }
   }
-}
+};
